# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Green Atlas API",
    "description": "Contract-first modular API for the deterministic greening prototype.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/agent-runs": {
      "post": {
        "tags": [
          "Autonomous agent"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_projects__project_id__agent_runs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/agent-runs/{run_id}": {
      "get": {
        "tags": [
          "Autonomous agent"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_projects__project_id__agent_runs__run_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/agent-runs/{run_id}/approve": {
      "post": {
        "tags": [
          "Autonomous agent"
        ],
        "summary": "Approve Run",
        "operationId": "approve_run_api_projects__project_id__agent_runs__run_id__approve_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunApproval"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/agent-runs/{run_id}/answer": {
      "post": {
        "tags": [
          "Autonomous agent"
        ],
        "summary": "Answer Run",
        "description": "Add one user answer and resume the same durable autonomous run.",
        "operationId": "answer_run_api_projects__project_id__agent_runs__run_id__answer_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunAnswer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/agent-runs/{run_id}/run": {
      "post": {
        "tags": [
          "Autonomous agent"
        ],
        "summary": "Run Agent",
        "operationId": "run_agent_api_projects__project_id__agent_runs__run_id__run_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentIntent": {
        "properties": {
          "raw_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Raw Text"
          },
          "goal": {
            "$ref": "#/components/schemas/Goal"
          },
          "scope_mode": {
            "type": "string",
            "enum": [
              "explicit",
              "selection",
              "delegated",
              "project"
            ],
            "title": "Scope Mode",
            "default": "explicit"
          },
          "explicit_zone_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 80,
            "title": "Explicit Zone Ids"
          },
          "explicit_object_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5000,
            "title": "Explicit Object Ids"
          },
          "hard_constraints": {
            "items": {
              "$ref": "#/components/schemas/HardConstraint"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Hard Constraints"
          },
          "unresolved_requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Unresolved Requirements"
          },
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/Preference"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Preferences"
          },
          "delegations": {
            "items": {
              "$ref": "#/components/schemas/Delegation"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Delegations"
          },
          "evidence": {
            "$ref": "#/components/schemas/IntentEvidence"
          },
          "arrangement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "plant_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "tree",
                  "shrub",
                  "mixed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Kind"
          },
          "species_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Species Ids"
          },
          "post_action": {
            "anyOf": [
              {
                "type": "string",
                "const": "focus_map"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "raw_text",
          "goal"
        ],
        "title": "AgentIntent",
        "description": "Immutable user intent atoms plus explicit delegations.\n\n``constraints`` contains only requirements, while subjective words such as\n«плотно» belong in ``preferences``.  Normative policy is represented by a\npolicy-owned HardConstraint and is never generated as user text."
      },
      "AgentRunAnswer": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 800,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AgentRunAnswer"
      },
      "AgentRunApproval": {
        "properties": {
          "preview_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AgentRunApproval"
      },
      "AgentRunCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Text"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AgentRunCreate"
      },
      "AgentRunEvent": {
        "properties": {
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "kind": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "kind",
          "created_at"
        ],
        "title": "AgentRunEvent"
      },
      "AgentRunRecord": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/AgentRunState"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AgentRunEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRunRecord"
      },
      "AgentRunState": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Run Id"
          },
          "project_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Project Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "waiting_question",
              "waiting_approval",
              "waiting_job",
              "finished",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "default": "queued"
          },
          "intent": {
            "$ref": "#/components/schemas/AgentIntent"
          },
          "resolved_scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolvedScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_zone_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 80,
            "title": "Candidate Zone Ids"
          },
          "snapshot_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Snapshot Version"
          },
          "plan_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Version"
          },
          "step": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Step",
            "default": 0
          },
          "tool_calls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 512,
            "title": "Tool Calls"
          },
          "tool_fingerprints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 512,
            "title": "Tool Fingerprints"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 512,
            "title": "Evidence Refs"
          },
          "last_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_question": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Question"
          },
          "pending_approval": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Approval"
          },
          "outcome_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Ref"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolError"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_steps": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 64
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "project_id",
          "intent",
          "snapshot_version"
        ],
        "title": "AgentRunState",
        "description": "Durable state restored after a process crash or human interruption."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolvedScope": {
        "properties": {
          "project_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Project Id"
          },
          "zone_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 80,
            "title": "Zone Ids"
          },
          "object_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5000,
            "title": "Object Ids"
          },
          "basis": {
            "type": "string",
            "enum": [
              "user",
              "selection",
              "agent",
              "policy"
            ],
            "title": "Basis"
          },
          "criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Criteria"
          },
          "source_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Source Revision"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Evidence Refs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "basis",
          "source_revision"
        ],
        "title": "ResolvedScope",
        "description": "One authoritative scope chosen from explicit input or tool evidence."
      },
...
      "ToolError": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 800,
            "minLength": 1,
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "remedy": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Remedy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ToolError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}